requirements.txt
tensorflow
numpy
opencv-python
pillow
pytesseract
regex
